+++ nix/ascii-numbers-1.2.0.nix
@@ -3,10 +3,8 @@
 }:
 mkDerivation {
   pname = "ascii-numbers";
-  version = "1.1.0.2";
-  sha256 = "dbfe4acb0eaa2f0c4ce68c6ac6072297cc485b5c7f9b3ad59375c36133b61837";
-  revision = "2";
-  editedCabalFile = "19x9mh11pb7j4ykf9vicprn6mlhcb9gwsk82gh5yk366k4r172d7";
+  version = "1.2.0.0";
+  sha256 = "59fc670d7a271c3cebb1850ae2e5d6d304590e8a62f45b5f5510230bf9798214";
   libraryHaskellDepends = [
     ascii-case ascii-char ascii-superset base bytestring hashable text
   ];