+++ nix/ascii-superset-1.3.0.nix
@@ -3,8 +3,8 @@
 }:
 mkDerivation {
   pname = "ascii-superset";
-  version = "1.2.7.0";
-  sha256 = "7602b831250eab00bd7133c9e6b2688a69c2c8f218eec525f449aa94076f1d90";
+  version = "1.3.0.0";
+  sha256 = "6338923c26fe6ee010741b2936d871efd3ecd6273a1f619366a26355de944e33";
   libraryHaskellDepends = [
     ascii-case ascii-caseless ascii-char base bytestring hashable text
   ];