+++ nix/ascii-th-1.2.0.nix
@@ -6,6 +6,8 @@ mkDerivation {
   pname = "ascii-th";
   version = "1.2.0.0";
   sha256 = "6f41d2216e473011cc660d05429c1ce2a7c76bdb87cd303a816871974b3a661f";
+  revision = "1";
+  editedCabalFile = "1r6z6brkfahs9zifjhr7bpqblkiajcjknkgx2i57jrn5s3b97phk";
   libraryHaskellDepends = [
     ascii-case ascii-caseless ascii-char ascii-superset base
     template-haskell